update dependencies
- use the latest version of everything except `canvas`

`canvas` v2.x is only supported with `jsdom` v12+ which supports node 6+.
But `jest` depends on an old version of `jsdom` because they still support node v4 🤦🏻‍♂️.

the latest version of `canvas` includes "prebuilt" versions which don't require any deps to install and use.
`canvas` ~1.x needs to be compiled ...

older versions of `canvas-prebuilt` no longer appear to be available and/or have security warnings, so we need to use `canvas`.

We could use a custom jest environment to take advantage of the latest version of `jsdom`,
and there is an npm package for that: https://github.com/dmnsgn/jest-environment-jsdom-latest
- but it hasn't been updated to `jsdom` v12+, and it's too much work to fork and fix - the repo has security vulnerabilities itself.